Re-start of DEVONthink-friendly version of script
Attempt to minimize the amount of DT-specific code ran into trouble.
Re-starting that script with a complete block of commands for Finder
selections and DT selections.  The DT version will just ignore the file
items piped in from QS, instead of trying to recover DT items from path
information (which might not be possible).